add /auth/ page
